feat(#678): move tab detail pages
